async_rest_api_dynamodb/requirements.txt
fastapi
boto3
uvicorn
pytest
moto
aws-cdk-lib
constructs
python-dotenv